planner_api/requirements.txt
google-genai>=0.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
